backend/requirements.txt
# FastMCP and Core Dependencies
fastmcp>=2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.12.0
# sqlite3 is built-in with Python

# Authentication and Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
authlib>=1.2.1

# HTTP Client and Networking
httpx>=0.25.0
websockets>=11.0
aiohttp>=3.9.0

# Configuration and Environment
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Utilities
click>=8.1.0
rich>=13.7.0
loguru>=0.7.0
asyncio-mqtt>=0.11.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# MCP Protocol Dependencies
mcp>=1.6.0
jsonrpc-base>=2.2.0
jsonrpc-async>=2.1.0

# GitHub Integration
PyGithub>=1.59.0
gitpython>=3.1.0